refactor(Rooms) change Room module for better consistency with routes
