+++ src/modules/Rooms/actions.js
