+++ src/modules/Rooms/components/Room.js
