+++ src/modules/Rooms/components/Stats.js
