+++ src/modules/Rooms/index.js
@@ -1,0 +1,1 @@
+export routes from './routes';
