+++ src/modules/Rooms/reducer.js
