+++ src/modules/Rooms/routes.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Route } from 'react-router';
 
 import Room from './Room';
